% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_obj.R
\name{eval_obj}
\alias{eval_obj}
\title{Evaluates SummarizedExperiment Object provided as input}
\usage{
eval_obj(obj)
}
\arguments{
\item{obj}{A summarizedExperiment object}
}
\value{
It does not return anything but halts the execution of program if any
of the check statements fail.
}
\description{
This script runs checks on the object to verify presense of required data at 
its specified location in the object. If any of the checks fail, the program 
is terminated with error message. If all checks pass then the control is
transfered back to the calling statement.
}
\examples{
 
data(mn_data)
eval_obj(mn_data)

}
