% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{extendClusterSet}
\alias{extendClusterSet}
\title{Extend cluster set to nearest neighbors on cluster graph.}
\usage{
extendClusterSet(graph, initial_set, max_neighbor_distance = 2)
}
\arguments{
\item{graph}{Graph in igraph format generated by makeClusterGraph.}

\item{initial_set}{Vector of cluster labels}

\item{max_neighbor_distance}{Include more distantly related nodes
by performing neigbor extension max_neighbor_distance rounds.}
}
\value{
Character vector including initial cluster set and all
neighboring clusters (if any).
}
\description{
Note that the graph is directed, i.e. neighbors are retrieved
by following arrows that start from the initial clusters.
}
