% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set.R
\docType{data}
\name{gene_set}
\alias{gene_set}
\title{gene_set}
\format{\describe{
  \item{genesets}{List containing gene symbols for 10 GO function (GO:0016853
  , GO:0005615, GO:0005768, GO:0007067, GO:0065003, GO:0042592, GO:0005929,
  GO:0008565, GO:0016829, GO:0022857) downloaded from the Gene Ontology
  Consortuum August 2015 \url{http://www.geneontology.org/}}
}}
\source{
Dataset:\url{https://github.com/mm-shah/MetaNeighbor/tree/master/data}
}
\usage{
gene_set
}
\description{
List containing gene symbols for 10 GO function
}
\keyword{datasets}
