% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_hits.R
\name{get_top_hits}
\alias{get_top_hits}
\title{Find reciprocal top hits}
\usage{
get_top_hits(cell_NV, mn_data, threshold = 0.95, filename)
}
\arguments{
\item{cell_NV}{A matrix of celltype-to-celltype AUROC scores
(output from \code{\link{run_MetaNeighbor_US}})}

\item{mn_data}{A SummarizedExperiment object containing gene-by-sample
expression matrix.}

\item{threshold}{Default value 0.95. Must be between [0-1]}

\item{filename}{Set path and name of output file}
}
\value{
Creates a file and returns a dataframe with cell types that are
either reciprocal best matches, and/or those with AUROC values greater than
or equal to threshold value
}
\description{
Identifies reciprocal top hits and high scoring cell type pairs
}
\examples{
data(mn_data)
var_genes = get_variable_genes(mn_data)
celltype_NV = run_MetaNeighbor_US(var_genes, mn_data)
top_hits = get_top_hits(celltype_NV,
                        mn_data,
                        threshold=0.9,
                        filename="filename.txt")
top_hits

}
