% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_genes.R
\name{get_variable_genes}
\alias{get_variable_genes}
\title{Identify a highly variable gene set}
\usage{
get_variable_genes(mn_data)
}
\arguments{
\item{mn_data}{A SummarizedExperiment object containing gene-by-sample
expression matrix.}
}
\value{
The output is a vector of gene names that are highly variable in
every experiment (intersect)
}
\description{
Identifies genes with high variance compared to their median expression
(top quartile) within each experimentCertain function
}
\examples{
data(mn_data)
var_genes = get_variable_genes(mn_data)
var_genes

}
