% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOhuman.R
\docType{data}
\name{GOhuman}
\alias{GOhuman}
\title{GOhuman}
\format{\describe{
  \item{genesets}{List containing gene symbols for 71 GO function (GO slim terms containing between 50 and 1,000 genes) downloaded from the Gene Ontology
  Consortium August 2015 \url{http://www.geneontology.org/page/download-annotations}}
}}
\source{
Dataset: \url{https://github.com/mm-shah/MetaNeighbor/tree/master/data}
| Paper: \url{https://www.biorxiv.org/content/early/2017/06/16/150524}
}
\usage{
GOhuman
}
\description{
List containing gene symbols for 71 GO function
}
\keyword{datasets}
