% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{makeClusterName}
\alias{makeClusterName}
\title{Make cluster names in format 'study_id|cell_type'}
\usage{
makeClusterName(study_id, cell_type)
}
\arguments{
\item{study_id}{Character vector containing study ids.}

\item{cell_type}{Character vector containing cell type names}
}
\value{
Character vector containing cluster names in the format
 study_id|cell_type.
}
\description{
Make cluster names in format 'study_id|cell_type'
}
