% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{mergeSCE}
\alias{mergeSCE}
\title{Merge multiple SingleCellExperiment objects.}
\usage{
mergeSCE(sce_list)
}
\arguments{
\item{sce_list}{A *named* list, where values are SingleCellExperiment objects
and names are SingleCellExperiment objects.}
}
\value{
A SingleCellExperiment object containing the input datasets with the
following limitations: (i) only genes common to all datasets are kept,
(ii) only colData columns common to all datasets are kept,
(iii) only assays common to all datasets (i.e., having the same name)
are kept, (iv) all other slots (e.g., reducedDims or rowData) will
be ignored and left empty.
The SingleCellExperiment object contains a "study_id" column,
mapping each cell to its original dataset (names in "sce_list").
}
\description{
Merge multiple SingleCellExperiment objects.
}
