% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNeighbor_US_data.R
\docType{data}
\name{MetaNeighbor_US_data}
\alias{MetaNeighbor_US_data}
\title{MetaNeighbor_US_data}
\format{\describe{
  \item{data}{A gene-by-sample expression matrix consisting of 3157 rows
  (genes) and 1051 columns (samples)}
  \item{pheno}{A sample metadata table (1051x2), that lists the dataset and
  cell type for each sample with column names: Study_ID and Celltype}
}}
\source{
Dataset:\url{https://github.com/mm-shah/MetaNeighbor/tree/master/data}
}
\usage{
MetaNeighbor_US_data
}
\description{
A list containing a collection of following: a gene matrix,
cell type labels, experiment labels and sets of genes.
}
\keyword{datasets}
