% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_voting_LeaveOneExpOut.R
\name{neighbor_voting_LeaveOneExpOut}
\alias{neighbor_voting_LeaveOneExpOut}
\title{Runs the neighbor voting algorithm.}
\usage{
neighbor_voting_LeaveOneExpOut(exp_labels, cell_labels, network, means = TRUE)
}
\arguments{
\item{exp_labels}{A numerical vector that indicates the dataset source of
each sample}

\item{cell_labels}{A sample by cell type matrix that indicates the cell type
of each sample (0-1)}

\item{network}{A sample by sample adjacency matrix, ranked and standardized
between 0-1}

\item{means}{Default TRUE, determines output formatting}
}
\value{
If \code{means = TRUE} (default) a vector containing the mean of
AUROC values across cross-validation folds will be returned. If FALSE a list
is returned containing a cell type by dataset matrix of AUROC scores, for
each fold of cross-validation. Default is over-ridden when more than one cell
type is assessed.
}
\description{
The function performs cell type identity prediction based on 'guilt by
association' using cross validation. Performance is evaluated by calculating
the AUROC for each cell type.
}
\examples{
data(MetaNeighbor_sample_data.RData)
AUROC_scores = run_MetaNeighbor(data = MetaNeighbor_sample_data$data,
                                experiment_labels = MetaNeighbor_sample_data$exp.lab,
                                celltype_labels = MetaNeighbor_sample_data$cell.lab,
                                genesets = MetaNeighbor_sample_data$genesets,
                                file_ext = "filename")
AUROC_scores

}
