% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{orderCellTypes}
\alias{orderCellTypes}
\title{Order cell types based on AUROC similarity matrix.}
\usage{
orderCellTypes(M, na_value = 0)
}
\arguments{
\item{M}{A square AUROC matrix as returned by MetaNeighborUS.}

\item{na_value}{Replace NA values with this value (default is 0).}
}
\value{
A hierarchical clustering object as returned by stats::hclust.
}
\description{
Order cell types based on AUROC similarity matrix.
}
