% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MetaNeighbor.R
\name{run_MetaNeighbor}
\alias{run_MetaNeighbor}
\title{Runs MetaNeighbor}
\usage{
run_MetaNeighbor(data, experiment_labels, celltype_labels, genesets, file_ext)
}
\arguments{
\item{data}{A gene-by-sample expression matrix.}

\item{experiment_labels}{A numerical vector that indicates the source of each
sample.}

\item{celltype_labels}{A matrix that indicates the cell type of each sample.}

\item{genesets}{Gene sets of interest provided as a list of vectors.}

\item{file_ext}{Specify the desired path and file name for output file.}
}
\value{
There are three outputs of the method:
     i.) A vector of AUROC scores representing the mean for each gene set
     tested is returned directly
     (see \code{\link{neighbor_voting_LeaveOneExpOut}}).
    ii.) A list containing the AUROC score for each cell type across all
    folds of cross-validation can be found in the first output file.
    iii.) A matrix containing the means for each cell type across folds of
    cross-validaiton can be found in the second output file.
}
\description{
For each gene set of interest, the function builds a network of rank
correlations between all cells. Next,It builds a network of rank correlations
between all cells for a gene set. Next, the neighbor voting predictor
produces a weighted matrix of predicted labels by performing matrix
multiplication between the network and the binary vector indicating cell type
membership, then dividing each element by the null predictor (i.e., node
degree). That is, each cell is given a score equal to the fraction of its
neighbors (including itself), which are part of a given cell type. For
cross-validation, we permute through all possible combinations of
leave-one-dataset-out cross-validation, and we report how well we can recover
cells of the same type as area under the receiver operator characteristic
curve (AUROC). This is repeated for all folds of cross-validation, and the
mean AUROC across folds is reported. Calls
\code{\link{neighbor_voting_LeaveOneExpOut}}.
}
\examples{
data(MetaNeighbor_sample_data)
AUROC_scores = run_MetaNeighbor(data = MetaNeighbor_sample_data$data,
                                experiment_labels = MetaNeighbor_sample_data$exp.lab,
                                celltype_labels = MetaNeighbor_sample_data$cell.lab,
                                genesets = MetaNeighbor_sample_data$genesets,
                                file_ext = "filename")
hist(AUROC_scores,
     main = "Sst Nos",
     xlab = "AUROC Scores",
     breaks = 20,
     xlim = c(0,1))
abline(v = mean(AUROC_scores), col = "red", lty = 2, lwd = 2)

}
