% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_clusters.R
\name{scoreMetaClusters}
\alias{scoreMetaClusters}
\title{Summarize meta-cluster information in a table.}
\usage{
scoreMetaClusters(meta_clusters, best_hits, outlier_label = "outliers")
}
\arguments{
\item{meta_clusters}{Meta-cluster list generated by extractMetaClusters.}

\item{best_hits}{Matrix of AUROCs used to extract meta-clusters.}

\item{outlier_label}{Element of meta-cluster list containing outlier
clusters.}
}
\value{
A data.frame. Column "meta_cluster" contains meta-cluster names,
"clusters" lists the clusters belonging to each meta-cluster,
"n_studies" is the number of studies spanned by the meta-cluster,
"score" is the average similarity between meta-cluster members
(average AUROC, NAs are treated as 0).
}
\description{
Summarize meta-cluster information in a table.
}
