% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{splitClusters}
\alias{splitClusters}
\title{Split clusters according to symmetric AUROC similarity.}
\usage{
splitClusters(mn_scores, k)
}
\arguments{
\item{mn_scores}{A symmetric AUROC matrix as generated by MetaNeighborUS.}

\item{k}{The number of desired cluster sets.}
}
\value{
A list of cluster sets, each cluster set is a character vector
 containg cluster labels.
}
\description{
This function computes hierarchical clustering to group similar clusters,
interpreting the AUROC matrix as a similarity matrix, then uses a standard
tree cutting algorithm to obtain groups of similar clusters. Note that the
cluster hierarchy corresponds exactly to the dendrogram shown when using
the plotHeatmap function.
}
\seealso{
\code{\link{plotHeatmap}}
}
