% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{splitTestClusters}
\alias{splitTestClusters}
\title{Split test clusters according to AUROC similarity to train clusters.}
\usage{
splitTestClusters(mn_scores, k)
}
\arguments{
\item{mn_scores}{An AUROC matrix as generated by MetaNeighborUS, usually with
the "trained_model" option.}

\item{k}{The number of desired cluster sets.}
}
\value{
A list of cluster sets, each cluster set is a character vector
containg cluster labels.
}
\description{
This function computes hierarchical clustering to group similar test
clusters, using similarity to train clusters as features, then uses a
standard tree cutting algorithm to obtain groups of similar clusters.
Note that the cluster hierarchy does *not* correspond to the row ordering of
plotHeatmapPretrained function, which uses a different heuristic.
}
\seealso{
\code{\link{plotHeatmapPretrained}}
}
