% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{standardizeLabel}
\alias{standardizeLabel}
\title{Remove special characters ("|") from labels to avoid later conflicts}
\usage{
standardizeLabel(labels, replace = "|", with = ".")
}
\arguments{
\item{labels}{Character vector containing study ids or cell type names.}

\item{replace}{Special character to replace}

\item{with}{Character to use instead of special character}
}
\value{
Character vector with replaced special characters.
}
\description{
Remove special characters ("|") from labels to avoid later conflicts
}
