% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{subsetClusterGraph}
\alias{subsetClusterGraph}
\title{Subset cluster graph to clusters of interest.}
\usage{
subsetClusterGraph(graph, vertices)
}
\arguments{
\item{graph}{Graph in igraph format generated by makeClusterGraph.}

\item{vertices}{Vector of cluster labels}
}
\value{
Graph in igraph format, where nodes have been restricted
to clusters of interests.
}
\description{
Subset cluster graph to clusters of interest.
}
\seealso{
\code{\link{extendClusterSet}}
}
