% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_genes.R
\name{get_variable_genes}
\alias{get_variable_genes}
\title{Identify a highly variable gene set}
\usage{
get_variable_genes(data, pheno)
}
\arguments{
\item{data}{A gene-by-sample expression matrix.}

\item{pheno}{A sample metadata table, that lists the dataset and cell type
for each sample with column names "Study_ID" and "Celltype".}
}
\value{
The output is a vector of gene names that are highly variable in
every experiment (intersect)
}
\description{
Identifies genes with high variance compared to their median expression
(top quartile) within each experimentCertain function
}
\examples{
data(MetaNeighbor_US_data)
var_genes = get_variable_genes(MetaNeighbor_US_data$data,
                               MetaNeighbor_US_data$pheno)
var_genes

}
