% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getCellType}
\alias{getCellType}
\title{Return cell type from a label in format 'study_id|cell_type'}
\usage{
getCellType(cluster_name)
}
\arguments{
\item{cluster_name}{Character vector containing cluster names in the
format study_id|cell_type.}
}
\value{
Character vector containing all cell type names.
}
\description{
Return cell type from a label in format 'study_id|cell_type'
}
