% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getStudyId}
\alias{getStudyId}
\title{Return study ID from a label in format 'study_id|cell_type'}
\usage{
getStudyId(cluster_name)
}
\arguments{
\item{cluster_name}{Character vector containing cluster names in the
format study_id|cell_type.}
}
\value{
Character vector containing all study ids.
}
\description{
Return study ID from a label in format 'study_id|cell_type'
}
