% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNeighbor_US_data.R
\docType{data}
\name{MetaNeighbor_US_data}
\alias{MetaNeighbor_US_data}
\title{MetaNeighbor_US_data}
\format{\describe{
  \item{data}{A gene-by-sample expression matrix consisting of 3157 rows
  (genes) and 1051 columns (cell types)}
  \item{pheno}{A sample metadata table consisting 1051 entries, that lists
  the Study_ID and Cell-type for each sample}
  \item{celltypes}{A vector of length 39 that lists the names of all cell
  types}
  }}
\source{
Dataset:\url{https://github.com/mm-shah/MetaNeighbor/tree/master/data}
}
\usage{
MetaNeighbor_US_data
}
\description{
A dataset containing a collection of following: a gene matrix,
cell type labels, experiment labels and sets of genes.
}
\keyword{datasets}
